WIP: Add separate workflow.py and split docs into workflow development and deployment
