+++ .secrets.baseline
@@ -179,7 +179,7 @@
         "filename": "docs/docs/en/user-guide/getting-started/index.md",
         "hashed_secret": "3be2e9dd1980856faddf5bd205e3ff96b24776aa",
         "is_verified": false,
-        "line_number": 373,
+        "line_number": 445,
         "is_secret": false
       }
     ],
@@ -224,5 +224,5 @@
       }
     ]
   },
-  "generated_at": "2024-11-05T06:29:55Z"
+  "generated_at": "2024-11-05T09:47:10Z"
 }
